# server/api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: StuHelper API
  version: 1.0.0
  description: StuHelper 学习助手平台 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080
    description: Development
  - url: https://api.stuhelper.com
    description: Production
...
paths:
...
  /api/v1/course/review/admin/teachers:
    get:
      tags:
        - Admin
      operationId: listAdminTeachers
      summary: 获取教师列表（管理员）
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: departmentID
          in: query
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 教师分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - list
                          - total
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/AdminTeacher'
                          total:
                            type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
        - Admin
      operationId: createTeacher
      summary: 新增教师
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminTeacher'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/admin/teachers/{teacherID}:
    put:
      tags:
        - Admin
      operationId: updateTeacher
      summary: 更新教师
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeacherIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeacherRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
        - Admin
      operationId: deleteTeacher
      summary: 删除教师
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeacherIDPath'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/admin/sensitive-words:
    get:
      tags:
        - Admin
      operationId: listSensitiveWords
      summary: 获取敏感词列表
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum:
              - block
              - warn
              - review
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 敏感词分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - list
                          - total
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/AdminSensitiveWord'
                          total:
                            type: integer
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
        - Admin
      operationId: createSensitiveWord
      summary: 新增敏感词
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSensitiveWordRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminSensitiveWord'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/course/review/admin/sensitive-words/{sensitiveWordID}:
    put:
      tags:
        - Admin
      operationId: updateSensitiveWord
      summary: 更新敏感词
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SensitiveWordIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSensitiveWordRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
        - Admin
      operationId: deleteSensitiveWord
      summary: 删除敏感词
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SensitiveWordIDPath'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseBody'
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          enum:
            - true
    ErrorResponseBody:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: A00400
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    MessageData:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    AdminTeacher:
      type: object
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        departmentID:
          type:
            - integer
            - 'null'
          format: int64
        departmentName:
          type:
            - string
            - 'null'
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateTeacherRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        departmentID:
          type: integer
          format: int64
    UpdateTeacherRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        departmentID:
          type:
            - integer
            - 'null'
          format: int64
    AdminSensitiveWord:
      type: object
      required:
        - id
        - word
        - category
        - level
        - isActive
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        word:
          type: string
        category:
          type: string
        level:
          type: string
          enum:
            - block
            - warn
            - review
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateSensitiveWordRequest:
      type: object
      required:
        - word
      properties:
        word:
          type: string
          minLength: 1
          maxLength: 100
        category:
          type: string
          maxLength: 50
          default: general
        level:
          type: string
          enum:
            - block
            - warn
            - review
          default: block
    UpdateSensitiveWordRequest:
      type: object
      properties:
        word:
          type: string
          minLength: 1
          maxLength: 100
        category:
          type: string
          maxLength: 50
        level:
          type: string
          enum:
            - block
            - warn
            - review
        isActive:
          type: boolean
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: 每页数量
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    TeacherIDPath:
      name: teacherID
      in: path
      required: true
      description: 教师 ID
      schema:
        type: integer
        format: int64
...
    SensitiveWordIDPath:
      name: sensitiveWordID
      in: path
      required: true
      description: 敏感词 ID
      schema:
        type: string
        format: uuid